Add Pump Offline  Errors and Update Operating Mode
